streamlit_app/requirements.txt
numpy==1.21.2
pandas==1.3.3
streamlit==0.85.0
Pillow==8.3.2

tensorflow==2.8.0
keras==2.8.0
matplotlib==3.2.2
nibabel==3.0.2
python==3.7.13